Refactor collector module.
This module started to be a big too big, so it needed to be split.
